# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PCCraft Marketplace API
  version: 1.0.0
  description: REST API for the PCCraft Marketplace platform -- multi-vendor PC components
    e-commerce with recommendations and a data-driven compatibility engine.
paths:
...
  /api/v1/admin/products/:
    get:
      operationId: admin_products_retrieve
      description: '``GET /api/v1/admin/products/`` -- paginated admin product list.'
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
  /api/v1/admin/products/{product_id}/:
    get:
      operationId: admin_products_retrieve_2
      description: '``GET / DELETE /api/v1/admin/products/{id}/``.'
      parameters:
      - in: path
        name: product_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
    delete:
      operationId: admin_products_destroy
      description: '``GET / DELETE /api/v1/admin/products/{id}/``.'
      parameters:
      - in: path
        name: product_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/admin/products/{product_id}/moderate/:
    post:
      operationId: admin_products_moderate_create
      description: '``POST /api/v1/admin/products/{id}/moderate/`` -- status flip.'
      parameters:
      - in: path
        name: product_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
  /api/v1/products/:
    get:
      operationId: products_list
      description: '``GET /api/v1/products/`` and ``GET /api/v1/products/{slug}/``.'
      parameters:
      - in: query
        name: brand
        schema:
          type: array
          items:
            type: string
        description: Multiple values may be separated by commas.
        explode: false
        style: form
      - in: query
        name: category
        schema:
          type: array
          items:
            type: string
        description: Multiple values may be separated by commas.
        explode: false
        style: form
      - in: query
        name: featured
        schema:
          type: boolean
      - in: query
        name: in_stock
        schema:
          type: boolean
      - in: query
        name: max_price
        schema:
          type: number
      - in: query
        name: min_price
        schema:
          type: number
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - products
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProductListList'
          description: ''
...
  /api/v1/products/{slug}/track-view/:
    post:
      operationId: products_track_view_create
      description: |-
        Record a product-view event for the recommendations engine.

        Spec §7.3: ``POST /api/v1/products/{slug}/track-view/``. Accepts
        anonymous viewers -- the frontend supplies an ``X-Session-Key``
        header for session attribution when the user is not logged in.
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - products
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductListRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductListRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProductListRequest'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
          description: ''
  /api/v1/products/search/:
    get:
      operationId: products_search_retrieve
      description: Lightweight search — delegates to the queryset filter.
      tags:
      - products
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
          description: ''
  /api/v1/products/trending/:
    get:
      operationId: products_trending_retrieve
      description: |-
        Top sellers in the last 30 days, capped to 12.

        Falls back to ``-total_sold`` when no order data is available
        yet (the orders app is built in Module 4).
      tags:
      - products
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
          description: ''
...
  /api/v1/search/:
    get:
      operationId: search_retrieve
      description: Full-text + faceted product search (Module 11 §11.1).
      tags:
      - search
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          description: No response body
...
  /api/v1/search/trending/:
    get:
      operationId: search_trending_retrieve
      description: Top ``TRENDING_LIMIT`` most-frequent non-zero queries in the last
        7 days.
      tags:
      - search
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          description: No response body
...
  /api/v1/vendor/products/:
    get:
      operationId: vendor_products_list
      description: '``/api/v1/vendor/products/`` -- scoped to the requesting vendor.'
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - vendor
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProductWriteList'
          description: ''
    post:
      operationId: vendor_products_create
      description: '``/api/v1/vendor/products/`` -- scoped to the requesting vendor.'
      tags:
      - vendor
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductWrite'
          description: ''
...
  /api/v1/vendor/products/{slug}/images/:
    post:
      operationId: vendor_products_images_create
      description: '``/api/v1/vendor/products/`` -- scoped to the requesting vendor.'
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - vendor
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductWrite'
          description: ''
...
  /api/v1/vendor/products/{slug}/images/reorder/:
    post:
      operationId: vendor_products_images_reorder_create
      description: '``/api/v1/vendor/products/`` -- scoped to the requesting vendor.'
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - vendor
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductWrite'
          description: ''
...
components:
  schemas:
...
    PaginatedProductListList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/ProductList'
    PaginatedProductWriteList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/ProductWrite'
...
    ProductListRequest:
      type: object
      description: Used for grid/card views in the catalog.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          minLength: 1
          maxLength: 220
          pattern: ^[-a-zA-Z0-9_]+$
        short_description:
          type: string
          maxLength: 500
        base_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        discounted_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
          nullable: true
        stock_quantity:
          type: integer
          maximum: 2147483647
          minimum: 0
        average_rating:
          type: string
          format: decimal
          pattern: ^-?\d{0,1}(?:\.\d{0,2})?$
        review_count:
          type: integer
          maximum: 2147483647
          minimum: 0
        total_sold:
          type: integer
          maximum: 2147483647
          minimum: 0
        is_featured:
          type: boolean
        status:
          $ref: '#/components/schemas/Status729Enum'
      required:
      - base_price
      - name
      - slug
...
    Status729Enum:
      enum:
      - DRAFT
      - ACTIVE
      - PAUSED
      - ARCHIVED
      - HIDDEN
      type: string
      description: |-
        * `DRAFT` - Draft
        * `ACTIVE` - Active
        * `PAUSED` - Paused
        * `ARCHIVED` - Archived
        * `HIDDEN` - Hidden
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT